Pages added

+++ VetAppointment/VetAppointment.API/Dtos/DrugDto.cs
@@ -4,7 +4,7 @@ namespace VetAppointment.API.Dtos
     public class DrugDto
     {
         public string Name { get; set; }
-        public float Quantity { get; set; }
-        public float UnitPrice { get; set; }
+        public double Quantity { get; set; }
+        public double UnitPrice { get; set; }
     }
 }
