+++ VetAppointment/VetAppointment.API/Dtos/PrescribedDrugDto.cs
@@ -2,8 +2,8 @@
 {
     public class PrescribedDrugDto
     {
-        public float Quantity { get; set; }
-        public float TotalCost { get; set; }
+        public double Quantity { get; set; }
+        public double TotalCost { get; set; }
 
         public Guid DrugId { get; set; }
